refactor(homepage): homepage3->homepage, old homepage->homepage-classic (4.3)
