+++ DECISIONS.md
@@ -1,1 +1,4 @@
 Cross-reference: [`docs/decisions/MONETIZATION_POLICY_V1.md`](docs/decisions/MONETIZATION_POLICY_V1.md) — v1 15% commission policy — internal/investor-facing.
+
+## ADR: homepage = AI conversation flow (2026-06-11)
+features/homepage is the AI conversation landing (ex-homepage3, serves /). features/homepage-classic is the legacy form (serves /form). Remove homepage-classic when /form traffic ~0.
