+++ src/features/homepage-classic/components/homepage-auth-gate.tsx
