+++ src/features/homepage-classic/components/homepage-discovery.test.tsx
