+++ src/features/homepage-classic/components/homepage-discovery.tsx
