+++ src/features/homepage/components/slot-chips.tsx
