+++ src/features/homepage/lib/extraction.ts
